Expanded cli with graphs and data retrieval, numerous improvements
Added line graphs to the CLI.

Added data retrieval commands to the CLI.

Generalised resolve.py to work so long as the dimensions are given
in the metadata.

Now reads phono3py direction data as a 3x3 array rather than a 6x1
array for consistency with other codes.

Updated READMEs.

There are still bugs which I'll fix soon!